refactor: move hocs and hooks to app top level
