+++ app/components/Channels/ChannelCardListItem.js
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import { opacity } from 'styled-system'
 import { Box as BaseBox, Flex as BaseFlex } from 'rebass'
 import { Bar, Card, Heading, Panel, Text } from 'components/UI'
-import withEllipsis from 'components/withEllipsis'
+import { withEllipsis } from 'hocs'
 import ChannelData from './ChannelData'
 import ChannelCapacity from './ChannelCapacity'
 import ChannelMoreButton from './ChannelMoreButton'
