+++ app/components/Channels/ChannelSummaryListItem.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { Box, Flex } from 'rebass'
 import { Card, Heading, Text } from 'components/UI'
-import withEllipsis from 'components/withEllipsis'
+import { withEllipsis } from 'hocs'
 import ChannelCapacity from './ChannelCapacity'
 import ChannelMoreButton from './ChannelMoreButton'
 import ChannelStatus from './ChannelStatus'
