+++ app/components/UI/Input.js
@@ -8,7 +8,7 @@ import { styles } from 'styled-system'
 import system from '@rebass/components'
 import { Flex } from 'rebass'
 import Search from 'components/Icon/Search'
-import withRequiredValidation from 'components/withRequiredValidation'
+import { withRequiredValidation } from 'hocs'
 import Message from './Message'
 import Label from './Label'
 import Span from './Span'
