+++ app/components/UI/TextArea.js
@@ -7,7 +7,7 @@ import styled, { withTheme } from 'styled-components'
 import { styles } from 'styled-system'
 import system from '@rebass/components'
 import { Flex } from 'rebass'
-import withRequiredValidation from 'components/withRequiredValidation'
+import { withRequiredValidation } from 'hocs'
 import Message from './Message'
 import Label from './Label'
 import Span from './Span'
