--- app/components/withEllipsis/index.js
@@ -1,3 +1,0 @@
-import withEllipsis from './withEllipsis'
-
-export default withEllipsis
