--- app/components/withLoading/index.js
@@ -1,3 +1,0 @@
-import withLoading from './withLoading'
-
-export default withLoading
