--- app/components/withNumberInputMask/index.js
@@ -1,3 +1,0 @@
-import withNumberInputMask from './withNumberInputMask'
-
-export default withNumberInputMask
