--- app/components/withOpenDialog/index.js
@@ -1,3 +1,0 @@
-import WithOpenDialog from './withOpenDialog'
-
-export default WithOpenDialog
