--- app/components/withRequiredValidation/index.js
@@ -1,3 +1,0 @@
-import withRequiredValidation from './withRequiredValidation'
-
-export default withRequiredValidation
