+++ app/containers/ModalStack.js
@@ -5,7 +5,7 @@ import { animated, Transition } from 'react-spring'
 import styled from 'styled-components'
 import { closeModal, modalSelectors } from 'reducers/modal'
 import { Modal, ModalOverlayStyles } from 'components/UI'
-import { useOnKeydown } from 'components/Util/hooks'
+import { useOnKeydown } from 'hooks'
 import Pay from 'containers/Pay'
 import Request from 'containers/Request'
 import Channels from 'containers/Channels'
