+++ app/hocs/index.js
@@ -1,0 +1,5 @@
+export withEllipsis from './withEllipsis'
+export withLoading from './withLoading'
+export withNumberInputMask from './withNumberInputMask'
+export WithOpenDialog from './withOpenDialog'
+export withRequiredValidation from './withRequiredValidation'
