+++ app/hocs/withEllipsis.js
