+++ app/hocs/withOpenDialog.js
