+++ app/hooks/index.js
@@ -1,3 +1,3 @@
-export useCloseOnUnmount from './closeOnUnmount'
+export useCloseOnUnmount from './useCloseOnUnmount'
 export useOnKeydown from './useOnKeydown'
 export useDebounce from './useDebounce'
