+++ app/hooks/useCloseOnUnmount.js
